created QuestionGenerator and split functionality with WQS
